% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_code_remind_me.R
\name{leaflet_code_remind_me}
\alias{leaflet_code_remind_me}
\title{Print basic leaflet map template}
\usage{
leaflet_code_remind_me()
}
\value{
None (prints to console).
}
\description{
This function prints a basic leaflet map template that can be used as a starting point for building maps.
}
\examples{
leaflet_code_remind_me()

}
