% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_default_tiles.R
\name{leaflet_default_tiles}
\alias{leaflet_default_tiles}
\title{Default tiles for a leaflet map.}
\usage{
leaflet_default_tiles(object)
}
\arguments{
\item{object}{A blank leaflet object to which you want to add default tiles.}
}
\value{
A leaflet object with default map tiles added.
}
\description{
This function adds default map tiles to a leaflet object. It includes OpenStreetMap (OSM),
Esri, and CartoDB tiles as options.
}
\examples{
\dontrun{
library(mapview)
mapview::breweries \%>\%
  leaflet::leaflet() \%>\%
  leaflet::addCircleMarkers(color = "black", weight = 2, fillColor = "blue") \%>\%
  leaflet_default_tiles() \%>\%
  leaflet::addLayersControl(baseGroups = leaflet_default_tiles_index(),
                            options = leaflet::layersControlOptions(collapsed = FALSE))
}

}
