% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_default_tiles_index.R
\name{leaflet_default_tiles_index}
\alias{leaflet_default_tiles_index}
\title{Commonly used Leaflet tiles.}
\usage{
leaflet_default_tiles_index()
}
\value{
A character vector of commonly used Leaflet tile names.
}
\description{
This function returns a character vector containing the names of commonly used Leaflet
tiles, including "OSM (default)," "Esri," and "CartoDB."
}
\examples{
leaflet_default_tiles_index()

}
