% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_popup_tbl_pretty.R
\name{leaflet_popup_tbl_pretty}
\alias{leaflet_popup_tbl_pretty}
\title{Create a popup table with formatted content.}
\usage{
leaflet_popup_tbl_pretty(data)
}
\arguments{
\item{data}{A spatial object.}
}
\value{
A leaflet popup table.
}
\description{
This function takes a spatial object and creates a popup table with formatted content.
}
\examples{
\dontrun{
library(mapview)
mapview::breweries \%>\%
  leaflet::leaflet() \%>\%
  leaflet::addCircleMarkers(color = "black", weight = 2, fillColor = "blue",
                            popup = leaflet_popup_tbl_pretty(mapview::breweries))
}

}
