% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_locator.R
\name{link_locator}
\alias{link_locator}
\title{link_locator}
\usage{
link_locator(
  search_list,
  geo_data_known,
  geo_data_known_col,
  geo_data_unknown,
  geo_data_unknown_id_col,
  large_buffer = 10000,
  small_buffer = 30
)
}
\arguments{
\item{search_list}{list of known links to use to reference unknown links with - used to map over reference network with known links}

\item{geo_data_known}{an sf object depicting network of known links}

\item{geo_data_known_col}{name of column that contains names contained in search_list}

\item{geo_data_unknown}{an sf object depicting network of unknown links that you want identified and located}

\item{geo_data_unknown_id_col}{name of primary key column in unknown link sf object}

\item{large_buffer}{a numeric value - should be sufficiently large}

\item{small_buffer}{a numeric value - should be small to ensure only link of interest is geolocated}
}
\value{
an sf object of referenced links
}
\description{
A function that references/codes/identifies referenced network links using a known spatial network.
}
\examples{
#work in progress
}
