% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_honeycomb_counts.R
\name{make_honeycomb_counts}
\alias{make_honeycomb_counts}
\title{Spatially aggregate spatial data into counts via honeycombs or squares.}
\usage{
make_honeycomb_counts(
  data,
  map_back_crs = 4326,
  honey_crs = 2285,
  honey_dim = 10000,
  honey_type = F
)
}
\arguments{
\item{data}{an SF object}

\item{map_back_crs}{an integer indicating which crs for the aggregated counts to be returned in - default is `4326`}

\item{honey_crs}{an integer indicating which crs to use in aggregation, aggregation cannot be perfomed useing lat/long, must be ft or m - default is `2285`}

\item{honey_dim}{a integer (in honey_crs units) indicating size of honeycomb or square}

\item{honey_type}{a boolean indicating what type of honeycomb to use - honeycomb (`F`) or square (`T`)}
}
\value{
an SF object (multipolygon) with counts per honeycomb unit
}
\description{
Spatially aggregate spatial data into counts via honeycombs or squares.
}
\examples{
mapview::breweries \%>\%
 make_honeycomb_counts(honey_type = T) \%>\%
filter(count > 0) \%>\%
 mapview::mapview(zcol = "count")
}
