% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_leaflet.R
\name{quick_leaflet}
\alias{quick_leaflet}
\title{Make a quick leaflet plot.}
\usage{
quick_leaflet(data, markers = F, lines = F, polys = F)
}
\arguments{
\item{data}{A spatial object.}

\item{markers}{A boolean value indicating whether to display marker symbols on the map. The default value is false.}

\item{lines}{A boolean value indicating whether to display lines on the map. The default value is false.}

\item{polys}{A boolean value indicating whether to display polygons on the map. The default value is false.}
}
\value{
A leaflet map.
}
\description{
This function creates a leaflet map with various options for displaying spatial data.
}
\examples{
\dontrun{
mapview::breweries \%>\%
quick_leaflet(markers = TRUE)
}
}
