% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_to.R
\name{rescale_to}
\alias{rescale_to}
\title{Rescale a numeric column to a specified maximum value.}
\usage{
rescale_to(column, value)
}
\arguments{
\item{column}{A numeric vector or column that you want to rescale.}

\item{value}{The maximum value to which the column will be rescaled.}
}
\value{
A numeric vector with the values rescaled to the specified maximum value.
}
\description{
This function takes a numeric column and scales it so that the values in
the column are rescaled to a specified maximum value. It's a general-purpose
function, but it's commonly used to control aesthetic controls like line widths, point sizes, or color scales in Leaflet maps.
}
\examples{
\dontrun{
# Rescale a numeric column to have a maximum value of 10
counties <- tigris::tracts(state = "WA", county = "king")
counties <- counties \%>\%
  filter(ALAND < median(ALAND)) \%>\%
  mutate(ALAND_rscld = rescale_to(ALAND, 10))

# Create a map using mapview to visualize the rescaled data
sf::st_centroid(counties) \%>\% mapview(cex = "ALAND_rscld", z = "ALAND_rscld")

}
}
