% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_calculate_heading_dataframe.R
\name{sf_calculate_heading_dataframe}
\alias{sf_calculate_heading_dataframe}
\title{Calculate Bearing for Polylines in Spatial Data Frame}
\usage{
sf_calculate_heading_dataframe(
  df,
  startLatCol,
  startLonCol,
  endLatCol,
  endLonCol,
  floor_divide = 1
)
}
\arguments{
\item{df}{A Spatial Data Frame (SF) containing start and end coordinates for polylines.}

\item{startLatCol}{Column name for the start latitude coordinates in the SF.}

\item{startLonCol}{Column name for the start longitude coordinates in the SF.}

\item{endLatCol}{Column name for the end latitude coordinates in the SF.}

\item{endLonCol}{Column name for the end longitude coordinates in the SF.}

\item{floor_divide}{An optional parameter to divide the calculated bearings for simplicity. Default is 1.}
}
\value{
A modified SF with an additional "bearing" attribute representing the directionality of the polylines.
}
\description{
This function calculates the bearing or heading in degrees for polylines in a Spatial Data Frame (SF).
}
\examples{
\dontrun{
# Example usage
result_sf <- sf_calculate_heading_dataframe(input_sf, "start_lat", "start_lon", "end_lat", "end_lon", floor_divide = 5)
}

}
