% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_chck_colinear_bearing.R
\name{st_chck_colinear_bearing}
\alias{st_chck_colinear_bearing}
\title{Check Colinearity of Bearings}
\usage{
st_chck_colinear_bearing(df, bearing_col_1, bearing_col_2, error)
}
\arguments{
\item{df}{A data frame containing the bearing columns.}

\item{bearing_col_1}{A string specifying the name of the first bearing column.}

\item{bearing_col_2}{A string specifying the name of the second bearing column.}

\item{error}{A numeric value specifying the error margin (in degrees) for collinearity checks.}
}
\value{
A data frame with an added column `is_colinear`, indicating whether the bearings in each row are collinear within the specified error margin.
}
\description{
This function checks whether two columns of bearings (in degrees) in a data frame are collinear within a specified error margin. It can be used to evaluate the directional consistency of network link bearings.
}
\examples{
\dontrun{
data.frame(
  bearing_1 = c(0),
  bearing_2 = c(5, 6, 10, 11, 15, 355, 0)
) \%>\%
  st_chck_colinear_bearing(
    "bearing_1", "bearing_2", error = 5
  )
}
}
