% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_line_sample_to_points.R
\name{st_line_sample_to_points}
\alias{st_line_sample_to_points}
\title{Convert LineString to Sampled Points}
\usage{
st_line_sample_to_points(sf_object, samp_dist = 100, crs)
}
\arguments{
\item{sf_object}{A spatial object of class 'sf' representing LineString geometry.}

\item{samp_dist}{The sampling distance, specifying the distance between consecutive sampled points along the LineString. The default value is 100.}

\item{crs}{The Coordinate Reference System (CRS) for the LineString geometry. If not specified, the function assumes the CRS of the input sf_object.}
}
\value{
A spatial object of class 'sf' representing the sampled points along the LineString.
}
\description{
Converts a LineString spatial object into a set of sampled points along the LineString.
}
\details{
The `st_line_sample_to_points` function takes a LineString spatial object and samples points along the LineString at a specified distance (`samp_dist`). The function first converts the input `sf_object` to LineString geometry by using the `st_cast` function and assigns a unique ID (`linestring_id`) to each LineString. It also creates a merge ID (`merge_id`) that corresponds to each LineString.

Next, the function samples points along each LineString using the `st_line_sample` function. The density of sampling is determined by the inverse of the `samp_dist` parameter. The resulting sampled points are then transformed to EPSG code 4326 (WGS84) and converted to a spatial object of class 'sf'.

Additional attributes, such as the `shape_id` from the original `sf_object_linestring`, are added to the sampled points using the `bind_cols` function. The resulting spatial object is then converted to Point geometry using the `st_cast` function and assigns an index (`index`) to each point. Finally, the function uses the `gauntlet::st_extract_coords` function to extract the longitude and latitude coordinates of the points, renaming them as `samp_lon` and `samp_lat`.

This function simplifies the process of sampling points along LineString geometries, which can be useful for various spatial analysis and visualization tasks.
}
\examples{
\dontrun{
# Convert LineString to sampled points
sampled_points <- st_line_sample_to_points(line_string_data, samp_dist = 100)

}
}
