% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_make_bounding_box.R
\name{st_make_bounding_box}
\alias{st_make_bounding_box}
\title{Make a bounding box of an sf object}
\usage{
st_make_bounding_box(base_geo)
}
\arguments{
\item{base_geo}{an sf object - can be a singular ploygon or multiple polygons.}
}
\value{
a dataframe of coordinates describing extent of supplied sf object
}
\description{
Make a bounding box of an sf object
}
\examples{
#none
}
