% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_self_buffer.R
\name{st_self_buffer}
\alias{st_self_buffer}
\title{Make a doughnut polygon around a spatial object.}
\usage{
st_self_buffer(spatial_object, rad_bg = NA, rad_sm = NA, nm = NULL)
}
\arguments{
\item{rad_bg}{Numeric. The radius for the background buffer.}

\item{rad_sm}{Numeric. The radius for the smaller buffer.}

\item{nm}{A string indicating the name of the buffer. Defaults to NULL.}

\item{object}{A spatial object of class 'sf', 'sfc' or 'sfg'.}
}
\value{
A list with two elements: a background buffer and a difference buffer.
}
\description{
Use an object and make a buffer around itself. Works well if you want to isolate an object but can only do it spatially.
}
\examples{
\dontrun{
curved_line = st_sfc(
  st_linestring(
    rbind(
      c(-122.68, 45.52)
      ,c(-122.63, 45.55)
      ,c(-122.58, 45.57)
      ,c(-122.55, 45.54)
    ))
  ,crs = 4326) \%>\%
  st_as_sf()

curved_line \%>\%
  st_self_buffer(rad_bg = 1000, rad_sm = 50, nm = "arbitary_name") \%>\%
  .[[2]] \%>\%
  mapview::mapview()

}
}
