% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extract_coords.R
\name{st_extract_coords}
\alias{st_extract_coords}
\title{Extract the longitude and latitude coordinates of a spatial object}
\usage{
st_extract_coords(spatial_object)
}
\arguments{
\item{spatial_object}{A spatial object representing points}
}
\value{
A data frame with two columns, "lon" for longitude and "lat" for latitude
}
\description{
This function extracts the longitude and latitude coordinates of a spatial object
representing points.
}
\examples{
\dontrun{
mapview::breweries \%>\%
 st_extract_coords()

}
}
